Working difference images
Two working Python scripts.

The absdiff script takes the first frame of the video and compares all subsequent frames to it and takes the difference. This isn't very effective for a live, lighting changing, shadowy, webcam like I need.

main.py uses CV's background subtractor along with a threshold and gaussian blur effect to make it a smoother detection.